+++ packages/payment-integration-api/src/checkout-buttons/default-checkout-button-height.ts
@@ -1,3 +1,3 @@
-const DefaultCheckoutButtonHeight = 36;
+const DefaultCheckoutButtonHeight = 40;
 
 export default DefaultCheckoutButtonHeight;