+++ packages/paypal-commerce-integration/src/paypal-commerce-credit-card/create-paypal-commerce-credit-cards-payment-strategy.ts
@@ -3,9 +3,9 @@ import {
     toResolvableModule,
 } from '@bigcommerce/checkout-sdk/payment-integration-api';
 import {
-    createPayPalCommerceFastlaneUtils,
-    createPayPalCommerceSdk,
-} from '@bigcommerce/checkout-sdk/paypal-commerce-utils';
+    createPayPalFastlaneUtils,
+    createPayPalSdkScriptLoader,
+} from '@bigcommerce/checkout-sdk/paypal-utils';
 
 import createPayPalCommerceIntegrationService from '../create-paypal-commerce-integration-service';
 
@@ -17,8 +17,8 @@ const createPaypalCommerceCreditCardsPaymentStrategy: PaymentStrategyFactory<
     new PayPalCommerceCreditCardsPaymentStrategy(
         paymentIntegrationService,
         createPayPalCommerceIntegrationService(paymentIntegrationService),
-        createPayPalCommerceSdk(),
-        createPayPalCommerceFastlaneUtils(),
+        createPayPalSdkScriptLoader(),
+        createPayPalFastlaneUtils(),
     );
 
 export default toResolvableModule(createPaypalCommerceCreditCardsPaymentStrategy, [