+++ packages/paypal-utils/src/create-paypal-sdk-script-loader.ts
@@ -1,0 +1,7 @@
+import { createScriptLoader } from '@bigcommerce/script-loader';
+
+import PayPalSdkScriptLoader from './paypal-sdk-script-loader';
+
+export default function createPayPalSdkScriptLoader(): PayPalSdkScriptLoader {
+    return new PayPalSdkScriptLoader(createScriptLoader());
+}