+++ packages/paypal-utils/src/index.ts
@@ -10,8 +10,8 @@ export { default as isPaypalProviderError } from './utils/is-paypal-provider-err
  * PayPal Sdk exports
  *
  * */
-export { default as createPaypalSdk } from './create-paypal-sdk';
-export { default as PaypalSdk } from './paypal-sdk-script-loader';
+export { default as createPayPalSdkScriptLoader } from './create-paypal-sdk-script-loader';
+export { default as PayPalSdkScriptLoader } from './paypal-sdk-script-loader';
 
 /**
  *